Refactor all the pages files to use SASS
