+++ pages/auth/index.module.scss
@@ -5,9 +5,10 @@
   flex-direction: column;
   margin: 100px auto;
   width: 50%;
+  
+  .signUpLink {
+    margin-top: 10px;
+    font-size: 20px;
+  }
 }
 
-.signUpLink {
-  margin-top: 10px;
-  font-size: 20px;
-}
