+++ pages/sign-up/index.module.scss
@@ -4,9 +4,9 @@
   flex-direction: column;
   margin: 100px auto;
   width: 50%;
-}
-
-.loginLink {
-  margin-top: 10px;
-  font-size: 20px;
+  
+  .loginLink {
+    margin-top: 10px;
+    font-size: 20px;
+  }
 }
